feat: переработать архитектуру в части хранения пользователей, исправить неработающие тесты
