+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -1,0 +1,10 @@
+package ru.yandex.practicum.filmorate.storage.user;
+
+import ru.yandex.practicum.filmorate.model.User;
+import java.util.Collection;
+
+public interface UserStorage {
+    Collection<User> findAllUsers();
+    User createUser(User user);
+    User updateUser(User user);
+}
